Move the atlasscript stuff back to atlastool.cpp
